Security-Enforcement-Agent/requirements.txt
kafka-python
requests
schedule
prometheus_client
flask